requirements.txt
python-telegram-bot==20.0
requests
icalendar
python-dotenv
aiohttp>=3.8.0
